+++ app/routes/logout.tsx
@@ -1,5 +1,5 @@
 import type { ActionFunctionArgs } from "@remix-run/node";
-import { logout } from "~/services/session.server";
+import { logout } from "~/.server/session";
 import {redirect} from "@remix-run/node";
 
 export async function action({ request }: ActionFunctionArgs) {
